pyproject.toml
[project]
name = "formula2video"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "edge-tts>=7.2.7",
    "manim>=0.19.1",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
    "pyyaml>=6.0.0",
    "mutagen>=1.47.0",
]
